Straighten dependencies
